added: multiprocess support, ao8 module
When Axiobus initializes, it now checks if the bus is active, if so, it just attaches and does not re-initialize.  Configuration calls now signal other processes using the library that a config change has happened, and they will read the needed module/channel configuration.  A change can be detected by calling the boolean "configurationChanged()" function on a module or channel.